refactor: organize services into logical subdirectories
- Group backup-related services into src/services/backup/
- Group notification-related services into src/services/notifications/
- Keep standalone services (errorLogger, locationService, toastService) at root level
- Update all imports across the entire codebase (40+ files affected)
- Fix relative imports in moved service files
- Update test mocks and dynamic imports
- Fix require statements in stores and integration tests

## Files Moved

### Backup Services (4 files)
- backupService.ts → backup/backupService.ts
- BackupServiceImpl.ts → backup/BackupServiceImpl.ts
- backupUtils.ts → backup/backupUtils.ts
- RestoreService.ts → backup/RestoreService.ts

### Notification Services (3 files)
- notificationService.ts → notifications/notificationService.ts
- notificationUtils.ts → notifications/notificationUtils.ts
- dailyCheckinService.ts → notifications/dailyCheckinService.ts

## Benefits

- Better organization with clear separation of concerns
- Easier navigation and maintenance
- Follows same successful pattern as screen refactoring
- All tests pass with no functionality changes
