+++ app/App.tsx
@@ -6,8 +6,8 @@ import Toast from 'react-native-toast-message';
 import AppNavigator from './src/navigation/AppNavigator';
 import { ThemeProvider } from './src/theme';
 import { getDatabase } from './src/database/db';
-import { notificationService } from './src/services/notificationService';
-import { dailyCheckinService } from './src/services/dailyCheckinService';
+import { notificationService } from './src/services/notifications/notificationService';
+import { dailyCheckinService } from './src/services/notifications/dailyCheckinService';
 import { logger } from './src/utils/logger';
 import { performanceMonitor } from './src/utils/performance';
 import ErrorBoundary from './src/components/ErrorBoundary';
