+++ app/src/screens/__tests__/SettingsScreen.test.tsx
@@ -4,17 +4,17 @@ import { Alert } from 'react-native';
 import SettingsScreen from '../settings/SettingsScreen';
 import { renderWithProviders } from '../../utils/screenTestHelpers';
 import { errorLogger } from '../../services/errorLogger';
-import { notificationService } from '../../services/notificationService';
+import { notificationService } from '../../services/notifications/notificationService';
 import { locationService } from '../../services/locationService';
-import { backupService } from '../../services/backupService';
+import { backupService } from '../../services/backup/backupService';
 import * as SQLite from 'expo-sqlite';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 jest.mock('../../services/errorLogger');
-jest.mock('../../services/notificationService');
+jest.mock('../../services/notifications/notificationService');
 jest.mock('../../services/locationService');
-jest.mock('../../services/dailyCheckinService');
-jest.mock('../../services/backupService');
+jest.mock('../../services/notifications/dailyCheckinService');
+jest.mock('../../services/backup/backupService');
 jest.mock('expo-sqlite');
 jest.mock('expo-notifications');
 jest.mock('@react-native-async-storage/async-storage');
