+++ app/src/screens/settings/NotificationSettingsScreen.tsx
@@ -15,8 +15,8 @@ import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { RootStackParamList } from '../../navigation/types';
 import { Ionicons } from '@expo/vector-icons';
 import { useTheme, ThemeColors } from '../../theme';
-import { notificationService, NotificationPermissions } from '../../services/notificationService';
-import { dailyCheckinService } from '../../services/dailyCheckinService';
+import { notificationService, NotificationPermissions } from '../../services/notifications/notificationService';
+import { dailyCheckinService } from '../../services/notifications/dailyCheckinService';
 import { useDailyCheckinSettingsStore } from '../../store/dailyCheckinSettingsStore';
 import DateTimePicker, { DateTimePickerEvent } from '@react-native-community/datetimepicker';
 import NotificationSettings from '../../components/NotificationSettings';
