+++ app/src/screens/settings/SettingsScreen.tsx
@@ -17,8 +17,8 @@ import { Ionicons } from '@expo/vector-icons';
 import { useTheme, ThemeColors } from '../../theme';
 import { buildInfo } from '../../buildInfo';
 import { errorLogger, ErrorLog } from '../../services/errorLogger';
-import { notificationService, NotificationPermissions } from '../../services/notificationService';
-import { dailyCheckinService } from '../../services/dailyCheckinService';
+import { notificationService, NotificationPermissions } from '../../services/notifications/notificationService';
+import { dailyCheckinService } from '../../services/notifications/dailyCheckinService';
 import { locationService } from '../../services/locationService';
 import { useDailyCheckinSettingsStore } from '../../store/dailyCheckinSettingsStore';
 import * as SQLite from 'expo-sqlite';
