+++ app/src/services/notifications/dailyCheckinService.ts
@@ -1,8 +1,8 @@
 import * as Notifications from 'expo-notifications';
-import { logger } from '../utils/logger';
-import { useDailyCheckinSettingsStore } from '../store/dailyCheckinSettingsStore';
-import { useDailyStatusStore } from '../store/dailyStatusStore';
-import { useEpisodeStore } from '../store/episodeStore';
+import { logger } from '../../utils/logger';
+import { useDailyCheckinSettingsStore } from '../../store/dailyCheckinSettingsStore';
+import { useDailyStatusStore } from '../../store/dailyStatusStore';
+import { useEpisodeStore } from '../../store/episodeStore';
 import { format } from 'date-fns';
 import { areNotificationsGloballyEnabled } from './notificationUtils';
